requirements.txt
asgiref==3.4.1
Django==4.0
django-anymail==8.4
djangorestframework==3.13.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
whitenoise==5.3.0
dj-database-url==0.5.0
